+++ src/components/ReplInput.svelte
@@ -1,5 +1,5 @@
 <script>
-	import Module from './Module.svelte';
+	import Module from './ReplInputModule.svelte';
 
 	export let examples = [];
 	export let selectedExample = null;
@@ -10,12 +10,12 @@
 	let modulesRef;
 
 	function removeModule(index) {
-		modules.splice(index, 1);
+		modules.splice(Input, 1);
 		modules = modules;
 	}
 
 	function toggleEntryModule(index) {
-		modules[index].isEntry = !modules[index].isEntry;
+		modules[Input].isEntry = !modules[Input].isEntry;
 	}
 
 	function createModule() {
@@ -37,6 +37,36 @@
 		selectedExample = null;
 	}
 </script>
+
+<header class="start-here clearfix">
+	<select bind:value="{selectedExample}">
+		<option disabled selected value="{null}">Select an example...</option>
+		{#each examples as example}
+			<option value="{example.id}">{example.title}</option>
+		{/each}
+	</select>
+
+	<button class="start-over" on:click="{clear}">Start over</button>
+</header>
+
+<div bind:this="{modulesRef}" class="modules">
+	{#each modules as module, i}
+		<Module
+			bind:name="{module.name}"
+			bind:code="{module.code}"
+			main="{i === 0}"
+			isEntry="{codeSplitting && module.isEntry}"
+			codeSplitting="{codeSplitting}"
+			on:remove="{() => removeModule(i)}"
+			on:toggle-entry="{() => toggleEntryModule(i)}"
+		/>
+	{/each}
+</div>
+
+<button class="new-module" on:click="{createModule}">
+	<span class="icon icon-plus"></span>
+	add module
+</button>
 
 <style>
 	select {
@@ -70,32 +100,3 @@
 		}
 	}
 </style>
-
-<header class="start-here clearfix">
-	<select bind:value="{selectedExample}">
-		<option disabled selected value="{null}">Select an example...</option>
-		{#each examples as example}
-			<option value="{example.id}">{example.title}</option>
-		{/each}
-	</select>
-
-	<button class="start-over" on:click="{clear}">Start over</button>
-</header>
-
-<div bind:this="{modulesRef}" class="modules">
-	{#each modules as module, i}
-		<Module
-			bind:name="{module.name}"
-			bind:code="{module.code}"
-			main="{i === 0}"
-			isEntry="{codeSplitting && module.isEntry}"
-			{codeSplitting}
-			on:remove="{() => removeModule(i)}"
-			on:toggle-entry="{() => toggleEntryModule(i)}" />
-	{/each}
-</div>
-
-<button class="new-module" on:click="{createModule}">
-	<span class="icon icon-plus"></span>
-	add module
-</button>
