+++ src/routes/guide/[lang]/_data.js
@@ -1,4 +1,4 @@
-import createGuide from './_createGuide.js';
+import createGuide from '../../../helpers/createGuide.js';
 
 const guide = createGuide();
 export default () => guide;
